api/requirements.txt
MODULE REQUIREMENT
tensorflow 
fastapi
uvicorn
python-multipart
pillow
tensorflow serving api
matplotlib
numpy